# openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /role:
        post:
            tags:
                - Role
            operationId: Role_Create
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/auth.v1.CreateRoleRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/auth.v1.RoleInfo'
    /role/list:
        get:
            tags:
                - Role
            operationId: Role_PageList
            parameters:
                - name: page.pageIndex
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: page.pageSize
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/auth.v1.PageRoleResponse'
    /role/{id}:
        get:
            tags:
                - Role
            operationId: Role_Get
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/auth.v1.RoleInfo'
        put:
            tags:
                - Role
            operationId: Role_Update
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/auth.v1.UpdateRoleRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
        delete:
            tags:
                - Role
            operationId: Role_Delete
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content: {}
...
components:
    schemas:
...
        auth.v1.CreateRoleRequest:
            type: object
            properties:
                code:
                    type: string
                name:
                    type: string
...
        auth.v1.PageRoleResponse:
            type: object
            properties:
                total:
                    type: string
                roles:
                    type: array
                    items:
                        $ref: '#/components/schemas/auth.v1.RoleInfo'
...
        auth.v1.RoleInfo:
            type: object
            properties:
                id:
                    type: string
                code:
                    type: string
                name:
                    type: string
                createdAt:
                    type: string
                updatedAt:
                    type: string
            description: 角色相关消息
...
        auth.v1.UpdateRoleRequest:
            type: object
            properties:
                id:
                    type: string
                code:
                    type: string
                name:
                    type: string
...